Use the dynamic extend avoidance dodge in all cases
This branch uses a version of the dynamic extend avoidance dodge as a
quick and dirty test of what like would look like in a steady state
without any object extensions.  Basically, instead of extending
descendants of Runtime::SyntaxNode we replace them with like-populated
objects of classes that descend from their class but include the
requested module, and rather than creating a new class each time we
memoize the finite (typically small) set of classes and reuse them.

Running the benchmark, we see new classes being created for the first
parse or two, and then no more.  Unfortunately, we also see a net
slowdown, as the overhead for the casting mechanism is too high to use
unless we must.